+++ Data/HashSet.hs
@@ -97,6 +97,10 @@ import qualified GHC.Exts as Exts
 import Data.Functor.Classes
 #endif
 
+#if MIN_VERSION_hashable(1,2,5)
+import qualified Data.Hashable.Lifted as H
+#endif
+
 -- | A set of values.  A set cannot contain duplicate values.
 newtype HashSet a = HashSet {
       asMap :: HashMap a ()
@@ -165,6 +169,11 @@ instance (Data a, Eq a, Hashable a) => Data (HashSet a) where
         _ -> error "gunfold"
     dataTypeOf _   = hashSetDataType
     dataCast1 f    = gcast1 f
+
+#if MIN_VERSION_hashable(1,2,6)
+instance H.Hashable1 HashSet where
+    liftHashWithSalt h s = H.liftHashWithSalt2 h hashWithSalt s . asMap
+#endif
 
 instance (Hashable a) => Hashable (HashSet a) where
     hashWithSalt salt = hashWithSalt salt . asMap