+++ CHANGES.md
@@ -4,6 +4,8 @@
 
  * `Eq (HashSet a)` doesn't require `Hashable a` anymore, only `Eq a`.
 
+ * Add `Hashable1/2` with `hashable-1.2.6.0`
+
 ## 0.2.7.2
 
  * Don't use -fregs-graphs